Refactor lemma task to use new file structure.
