+++ spacy_llm/tasks/lemma/task.py
@@ -1,84 +1,20 @@
 from typing import Any, Callable, Dict, Iterable, List, Optional, Type
 
 import jinja2
-from pydantic import BaseModel
 from spacy.language import Language
 from spacy.scorer import Scorer
 from spacy.tokens import Doc
 from spacy.training import Example
 
-from ..registry import registry
-from ..ty import ExamplesConfigType, TaskResponseParser
-from .templates import read_template
-from .util import SerializableTask
-
-_DEFAULT_LEMMA_TEMPLATE_V1 = read_template("lemma")
-
-
-class LemmaExample(BaseModel):
-    text: str
-    lemmas: List[Dict[str, str]]
-
-
-def parse_responses_v1(docs: Iterable[Doc], responses: Iterable[str]) -> Iterable[Doc]:
-    for doc, prompt_response in zip(docs, responses):
-        parsed_response = [
-            [pr_part.strip() for pr_part in pr.split(":")]
-            for pr in prompt_response.replace("Lemmatized text:", "")
-            .replace("'''", "")
-            .strip()
-            .split("\n")
-        ]
-        tokens = [token for token in doc]
-
-        # If numbers of tokens recognized by spaCy and returned by LLM don't match, we don't attempt a partial
-        # match.
-        if len(tokens) != len(parsed_response):
-            yield doc
-
-        # Assign lemmas.
-        for token, lemma_info in zip(tokens, parsed_response):
-            if len(lemma_info) > 0:
-                token.lemma_ = lemma_info[1]
-
-        yield doc
-
-
-@registry.llm_misc("spacy.LemmaParser.v1")
-def make_lemma_parser():
-    return parse_responses_v1
-
-
-@registry.llm_tasks("spacy.Lemma.v1")
-def make_lemma_task(
-    template: str = _DEFAULT_LEMMA_TEMPLATE_V1,
-    parse_responses: Optional[TaskResponseParser] = None,
-    examples: ExamplesConfigType = None,
-):
-    """Lemma.v1 task factory.
-
-    template (str): Prompt template passed to the model.
-    parse_responses (Optional[TaskResponseParser]): Callable for parsing LLM responses for this task.
-    examples (Optional[Callable[[], Iterable[Any]]]): Optional callable that
-        reads a file containing task examples for few-shot learning. If None is
-        passed, then zero-shot learning will be used.
-    """
-    raw_examples = examples() if callable(examples) else examples
-    lemma_examples = (
-        [LemmaExample(**eg) for eg in raw_examples] if raw_examples else None
-    )
-
-    return LemmaTask(
-        template=template,
-        parse_responses=parse_responses if parse_responses else parse_responses_v1,
-        examples=lemma_examples,
-    )
+from ...ty import TaskResponseParser
+from ..util import SerializableTask
+from .util import DEFAULT_LEMMA_TEMPLATE_V1, LemmaExample
 
 
 class LemmaTask(SerializableTask[LemmaExample]):
     def __init__(
         self,
-        template: str = _DEFAULT_LEMMA_TEMPLATE_V1,
+        template: str = DEFAULT_LEMMA_TEMPLATE_V1,
         parse_responses: Optional[TaskResponseParser] = None,
         examples: Optional[List[LemmaExample]] = None,
     ):
